refactor: renamed slices to be generic
